requirements.txt
python-dotenv
Flask
pymysql
mysql-connector-python
Flask-JWT-Extended
marshmallowow